sensors/image-embedding/requirements.txt
sentence-transformers>=3.0.0
transformers>=4.44.0,<4.46.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
einops>=0.7.0
timm>=0.9.0
requests>=2.31.0
